Add test case for uncertain promotion scenario
If our active source disconnects at the same time as we're trying to
promote a source that's third on our list, our request can be overridden by the
promotion logic in the source disconnect handler.

I'm not sure that's a bug worth caring about, but I figured checking in a test
case to start discussion couldn't hurt.